Reflected discussion on the call
