+++ nemo/collections/asr/parts/utils/speaker_utils.py
@@ -15,9 +15,10 @@
 import json
 import math
 import os
+import shutil
 from copy import deepcopy
 from functools import reduce
-from typing import List
+from typing import Tuple, Dict, Union, List
 
 import numpy as np
 import omegaconf
@@ -27,7 +28,7 @@ from pyannote.core import Annotation, Segment, Timeline
 from pyannote.metrics.diarization import DiarizationErrorRate
 from tqdm import tqdm
 
-from nemo.collections.asr.parts.utils.nmesc_clustering import COSclustering
+from nemo.collections.asr.parts.utils.nmesc_clustering import COSclustering, get_argmin_mat
 from nemo.utils import logging
 
 
@@ -45,6 +46,14 @@ def get_uniqname_from_filepath(filepath):
         return uniq_id
     else:
         raise TypeError("input must be filepath string")
+
+def get_uniq_id_from_manifest_line(line: str) -> str:
+    """
+    Retrieve `uniq_id` from the `audio_filepath` in a manifest line.
+    """
+    dic = json.loads(line.strip())
+    uniq_id = get_uniqname_from_filepath(dic['audio_filepath'])
+    return uniq_id
 
 
 def get_uniq_id_with_dur(meta, deci=3):
@@ -229,6 +238,30 @@ def get_embs_and_timestamps(multiscale_embeddings_and_timestamps, multiscale_arg
 
     return embs_and_timestamps
 
+def get_timestamps(multiscale_timestamps, multiscale_args_dict):
+    """
+    The timestamps in `multiscale_timestamps` dictionary are indexed by scale index.
+    This function rearranges the extracted speaker embedding and timestamps by unique ID to make the further processing more convenient.
+
+    Args:
+        multiscale_timestamps (dict):
+            Dictionary of timestamps for each scale.
+        multiscale_args_dict (dict):
+            Dictionary of scale information: window, shift and multiscale weights.
+
+    Returns:
+        timestamps_dict (dict)
+            A dictionary containing embeddings and timestamps of each scale, indexed by unique ID.
+    """
+    timestamps_dict = {uniq_id: {'scale_dict': {}} for uniq_id in multiscale_timestamps[0].keys()}
+    for scale_idx in sorted(multiscale_args_dict['scale_dict'].keys()):
+        time_stamps = multiscale_timestamps[scale_idx]
+        for uniq_id in time_stamps.keys():
+            timestamps_dict[uniq_id]['scale_dict'][scale_idx] = {
+                'time_stamps': time_stamps[uniq_id],
+            }
+
+    return timestamps_dict
 
 def get_contiguous_stamps(stamps):
     """
@@ -945,6 +978,331 @@ def get_subsegments(offset: float, window: float, shift: float, duration: float)
 
     return subsegments
 
+def get_scale_mapping_argmat(uniq_embs_and_timestamps: Dict[str, dict]) -> Dict[int, torch.Tensor]:
+    """
+    Calculate cosine similarity values among speaker embeddings for each scale then
+    apply multiscale weights to calculate the fused similarity matrix.
+
+    Args:
+        uniq_embs_and_timestamps: (dict)
+            The dictionary containing embeddings, timestamps and multiscale weights.
+            If uniq_embs_and_timestamps contains only one scale, single scale diarization
+            is performed.
+
+    Returns:
+        scale_mapping_argmat (dict)
+            Dictionary containing scale mapping information matrix for each scale.
+    """
+    scale_mapping_argmat = {}
+    uniq_scale_dict = uniq_embs_and_timestamps['scale_dict']
+    session_scale_mapping_dict = get_argmin_mat(uniq_scale_dict)
+    for scale_idx in sorted(uniq_scale_dict.keys()):
+        mapping_argmat = session_scale_mapping_dict[scale_idx]
+        scale_mapping_argmat[scale_idx] = mapping_argmat
+    return scale_mapping_argmat
+
+def get_overlap_stamps(cont_stamps: List[str], ovl_spk_idx: List[str]):
+    """
+    Generate timestamps that include overlap speech. Overlap-including timestamps are created based on the segments that are
+    created for clustering diarizer. Overlap speech is assigned to the existing speech segments in `cont_stamps`.
+
+    Args:
+        cont_stamps (list):
+            Non-overlapping (single speaker per segment) diarization output in string format.
+            Each line contains the start and end time of segments and corresponding speaker labels.
+        ovl_spk_idx (list):
+            List containing segment index of the estimated overlapped speech. The start and end of segments are based on the
+            single-speaker (i.e., non-overlap-aware) RTTM generation.
+    Returns:
+        total_ovl_cont_list (list):
+            Rendered diarization output in string format. Each line contains the start and end time of segments and
+            corresponding speaker labels. This format is identical to `cont_stamps`.
+    """
+    ovl_spk_cont_list = [[] for _ in range(len(ovl_spk_idx))]
+    for spk_idx in range(len(ovl_spk_idx)):
+        for idx, cont_a_line in enumerate(cont_stamps):
+            start, end, speaker = cont_a_line.split()
+            if idx in ovl_spk_idx[spk_idx]:
+                ovl_spk_cont_list[spk_idx].append(f"{start} {end} speaker_{spk_idx}")
+    total_ovl_cont_list = []
+    for ovl_cont_list in ovl_spk_cont_list:
+        if len(ovl_cont_list) > 0:
+            total_ovl_cont_list.extend(merge_stamps(ovl_cont_list))
+    return total_ovl_cont_list
+
+
+def get_adaptive_threshold(estimated_num_of_spks: int, min_threshold: float, overlap_infer_spk_limit: int):
+    """
+    This function controls the magnitude of the sigmoid threshold based on the estimated number of speakers. As the number of
+    speakers becomes larger, diarization error rate is very sensitive on overlap speech detection. This function linearly increases
+    the threshold in proportion to the estimated number of speakers so more confident overlap speech results are reflected when
+    the number of estimated speakers are relatively high.
+
+    Args:
+        estimated_num_of_spks (int):
+            Estimated number of speakers from the clustering result.
+        min_threshold (float):
+            Sigmoid threshold value from the config file. This threshold value is minimum threshold value when `estimated_num_of_spks=2`
+        overlap_infer_spk_limit (int):
+            If the `estimated_num_of_spks` is less then `overlap_infer_spk_limit`, overlap speech estimation is skipped.
+
+    Returns:
+        adaptive_threshold (float):
+            Threshold value that is scaled based on the `estimated_num_of_spks`.
+    """
+    adaptive_threshold = min_threshold - (estimated_num_of_spks - 2) * (min_threshold - 1) / (
+        overlap_infer_spk_limit - 2
+    )
+    return adaptive_threshold
+
+
+def generate_speaker_timestamps(
+    clus_labels: List[Union[float, int]], msdd_preds: List[torch.Tensor], **params
+) -> Tuple[List[str], List[str]]:
+    '''
+    Generate speaker timestamps from the segmentation information. If `use_clus_as_main=True`, use clustering result for main speaker
+    labels and use timestamps from the predicted sigmoid values. In this function, the main speaker labels in `maj_labels` exist for
+    every subsegment steps while overlap speaker labels in `ovl_labels` only exist for segments where overlap-speech is occuring.
+
+    Args:
+        clus_labels (list):
+            List containing integer-valued speaker clustering results.
+        msdd_preds (list):
+            List containing tensors of the predicted sigmoid values.
+            Each tensor has shape of: (Session length, estimated number of speakers).
+        params:
+            Parameters for generating RTTM output and evaluation. Parameters include:
+                infer_overlap (bool): If False, overlap-speech will not be detected.
+                use_clus_as_main (bool): Add overlap-speech detection from MSDD to clustering results. If False, only MSDD output
+                                         is used for constructing output RTTM files.
+                overlap_infer_spk_limit (int): Above this limit, overlap-speech detection is bypassed.
+                use_adaptive_thres (bool): Boolean that determines whehther to use adaptive_threshold depending on the estimated
+                                           number of speakers.
+                max_overlap_spks (int): Maximum number of overlap speakers detected. Default is 2.
+                threshold (float): Sigmoid threshold for MSDD output.
+
+    Returns:
+        maj_labels (list):
+            List containing string-formated single-speaker speech segment timestamps and corresponding speaker labels.
+            Example: [..., '551.685 552.77 speaker_1', '552.99 554.43 speaker_0', '554.97 558.19 speaker_0', ...]
+        ovl_labels (list):
+            List containing string-formated additional overlapping speech segment timestamps and corresponding speaker labels.
+            Note that `ovl_labels` includes only overlapping speech that is not included in `maj_labels`.
+            Example: [..., '152.495 152.745 speaker_1', '372.71 373.085 speaker_0', '554.97 555.885 speaker_1', ...]
+    '''
+    msdd_preds.squeeze(0)
+    estimated_num_of_spks = msdd_preds.shape[-1]
+    overlap_speaker_list = [[] for _ in range(estimated_num_of_spks)]
+    infer_overlap = estimated_num_of_spks < int(params['overlap_infer_spk_limit'])
+    main_speaker_lines = []
+    if params['use_adaptive_thres']:
+        threshold = get_adaptive_threshold(
+            estimated_num_of_spks, params['threshold'], params['overlap_infer_spk_limit']
+        )
+    else:
+        threshold = params['threshold']
+    for seg_idx, cluster_label in enumerate(clus_labels):
+        msdd_preds.squeeze(0)
+        spk_for_seg = (msdd_preds[0, seg_idx] > threshold).int().cpu().numpy().tolist()
+        sm_for_seg = msdd_preds[0, seg_idx].cpu().numpy()
+
+        if params['use_clus_as_main']:
+            main_spk_idx = int(cluster_label[2])
+        else:
+            main_spk_idx = np.argsort(msdd_preds[0, seg_idx].cpu().numpy())[::-1][0]
+
+        if sum(spk_for_seg) > 1 and infer_overlap:
+            idx_arr = np.argsort(sm_for_seg)[::-1]
+            for ovl_spk_idx in idx_arr[: params['max_overlap_spks']].tolist():
+                if ovl_spk_idx != int(main_spk_idx):
+                    overlap_speaker_list[ovl_spk_idx].append(seg_idx)
+        main_speaker_lines.append(f"{cluster_label[0]} {cluster_label[1]} speaker_{main_spk_idx}")
+    cont_stamps = get_contiguous_stamps(main_speaker_lines)
+    maj_labels = merge_stamps(cont_stamps)
+    ovl_labels = get_overlap_stamps(cont_stamps, overlap_speaker_list)
+    return maj_labels, ovl_labels
+
+
+def get_uniq_id_list_from_manifest(manifest_file: str):
+    """Retrieve `uniq_id` values from the given manifest_file and save the IDs to a list.
+    """
+    uniq_id_list = []
+    with open(manifest_file, 'r', encoding='utf-8') as manifest:
+        for i, line in enumerate(manifest.readlines()):
+            line = line.strip()
+            dic = json.loads(line)
+            uniq_id = get_uniqname_from_filepath(dic['audio_filepath'])
+            uniq_id_list.append(uniq_id)
+    return uniq_id_list
+
+
+def get_id_tup_dict(uniq_id_list: List[str], test_data_collection, preds_list: List[torch.Tensor]):
+    """
+    Create session-level dictionary containing data needed to construct RTTM diarization output.
+
+    Args:
+        uniq_id_list (list):
+            List containing the `uniq_id` values.
+        test_data_collection (collections.DiarizationLabelEntity):
+            Class instance that is containing session information such as targeted speaker indices, audio filepath and RTTM filepath.
+        preds_list (list):
+            List containing tensors of predicted sigmoid values.
+
+    Returns:
+        session_dict (dict):
+            Dictionary containing session-level target speakers data and predicted simoid values in tensor format.
+    """
+    session_dict = {x: [] for x in uniq_id_list}
+    for idx, line in enumerate(test_data_collection):
+        uniq_id = get_uniqname_from_filepath(line.audio_file)
+        session_dict[uniq_id].append([line.target_spks, preds_list[idx]])
+    return session_dict
+
+def prepare_split_data(manifest_filepath, _out_dir, multiscale_args_dict, global_rank):
+    """
+    This function is needed for preparing diarization training data for multiscale diarization decoder (MSDD).
+    Prepare multiscale timestamp data for training. Oracle VAD timestamps from RTTM files are used as VAD timestamps.
+    In this function, timestamps for embedding extraction are extracted without extracting the embedding vectors.
+
+    Args:
+        manifest_filepath (str):
+            Input manifest file for creating audio-to-RTTM mapping.
+        _out_dir (str):
+            Output directory where timestamp json files are saved.
+
+    Returns:
+        multiscale_args_dict (dict):
+            - Dictionary containing two types of arguments: multi-scale weights and subsegment timestamps for each data sample.
+            - Each data sample has two keys: `multiscale_weights` and `scale_dict`.
+                - `multiscale_weights` key contains a list containing multiscale weights.
+                - `scale_dict` is indexed by integer keys which are scale index.
+            - Each data sample is indexed by using the following naming convention: `<uniq_id>_<start time in ms>_<end time in ms>`
+                Example: `fe_03_00106_mixed_626310_642300`
+    """
+    speaker_dir = os.path.join(_out_dir, 'speaker_outputs')
+    
+    # Only if this is for the first run of modelPT instance, remove temp folders.
+    if global_rank == 0:
+        if os.path.exists(speaker_dir):
+            shutil.rmtree(speaker_dir)
+        os.makedirs(speaker_dir)
+    split_audio_rttm_map = audio_rttm_map(manifest_filepath, attach_dur=True)
+
+    # Speech Activity Detection part
+    _speaker_manifest_path = os.path.join(speaker_dir, f'oracle_vad_manifest.json')
+    logging.info(f"Extracting oracle VAD timestamps and saving at {speaker_dir}")
+    if not os.path.exists(_speaker_manifest_path):
+        write_rttm2manifest(split_audio_rttm_map, _speaker_manifest_path, include_uniq_id=True)
+
+    multiscale_timestamps_by_scale = {}
+
+    # Segmentation
+    for scale_idx, (window, shift) in multiscale_args_dict['scale_dict'].items():
+        subsegments_manifest_path = os.path.join(speaker_dir, f'subsegments_scale{scale_idx}.json')
+        if not os.path.exists(subsegments_manifest_path):
+            # Sub-segmentation for the current scale (scale_idx)
+            segments_manifest_to_subsegments_manifest(
+                segments_manifest_file=_speaker_manifest_path,
+                subsegments_manifest_file=subsegments_manifest_path,
+                window=window,
+                shift=shift,
+                include_uniq_id=True,
+            )
+            logging.info(
+                f"Subsegmentation for timestamp extracted for: scale-{scale_idx} at {subsegments_manifest_path}"
+            )
+        multiscale_timestamps = extract_timestamps(subsegments_manifest_path)
+        multiscale_timestamps_by_scale[scale_idx] = multiscale_timestamps
+
+    multiscale_timestamps_dict = get_timestamps(multiscale_timestamps_by_scale, multiscale_args_dict)
+    return multiscale_timestamps_dict
+
+def extract_timestamps(manifest_file: str):
+    """
+    This method extracts timestamps from segments passed through manifest_file. 
+
+    Args:
+        manifest_file (str):
+            Manifest file containing segmentation information.
+    Returns:
+        time_stamps (dict):
+            Dictionary containing lists of timestamps.
+    """
+    logging.info(f"Extracting timestamps from {manifest_file} for multiscale subsegmentation.")
+    time_stamps = {}
+    with open(manifest_file, 'r', encoding='utf-8') as manifest:
+        for i, line in enumerate(manifest.readlines()):
+            line = line.strip()
+            dic = json.loads(line)
+
+            uniq_name = dic['uniq_id']
+            if uniq_name not in time_stamps:
+                time_stamps[uniq_name] = []
+            start = dic['offset']
+            end = start + dic['duration']
+            stamp = '{:.3f} {:.3f} '.format(start, end)
+            time_stamps[uniq_name].append(stamp)
+    return time_stamps
+
+def make_rttm_with_overlap(
+    manifest_file_path: str,
+    clus_label_dict: Dict[str, List[Union[float, int]]],
+    msdd_preds: List[torch.Tensor],
+    **params,
+):
+    """
+    Create RTTM files that include detected overlap speech. Note that the effect of overlap detection is only
+    notable when RTTM files are evaluated with `ignore_overlap=False` option.
+
+    Args:
+        manifest_file_path (str):
+            Path to the input manifest file.
+        clus_label_dict (dict):
+            Dictionary containing subsegment timestamps in float type and cluster labels in integer type.
+            Indexed by `uniq_id` string.
+        msdd_preds (list):
+            List containing tensors of the predicted sigmoid values.
+            Each tensor has shape of: (Session length, estimated number of speakers).
+        params:
+            Parameters for generating RTTM output and evaluation. Parameters include:
+                infer_overlap (bool): If False, overlap-speech will not be detected.
+            See docstrings of `generate_speaker_timestamps` function for other variables in `params`.
+
+    Returns:
+        all_hypothesis (list):
+            List containing Pyannote's `Annotation` objects that are created from hypothesis RTTM outputs.
+        all_reference
+            List containing Pyannote's `Annotation` objects that are created from ground-truth RTTM outputs
+    """
+    AUDIO_RTTM_MAP = audio_rttm_map(manifest_file_path)
+    manifest_file_lengths_list = []
+    all_hypothesis, all_reference = [], []
+    no_references = False
+    with open(manifest_file_path, 'r', encoding='utf-8') as manifest:
+        for i, line in enumerate(manifest.readlines()):
+            uniq_id = get_uniq_id_from_manifest_line(line)
+            manifest_dic = AUDIO_RTTM_MAP[uniq_id]
+            clus_labels = clus_label_dict[uniq_id]
+            manifest_file_lengths_list.append(len(clus_labels))
+            maj_labels, ovl_labels = generate_speaker_timestamps(clus_labels, msdd_preds[i], **params)
+            if params['infer_overlap']:
+                hyp_labels = maj_labels + ovl_labels
+            else:
+                hyp_labels = maj_labels
+            hypothesis = labels_to_pyannote_object(hyp_labels, uniq_name=uniq_id)
+            if params['out_rttm_dir']:
+                hyp_labels = sorted(hyp_labels, key=lambda x: float(x.split()[0]))
+                labels_to_rttmfile(hyp_labels, uniq_id, params['out_rttm_dir'])
+            all_hypothesis.append([uniq_id, hypothesis])
+            rttm_file = manifest_dic.get('rttm_filepath', None)
+            if rttm_file is not None and os.path.exists(rttm_file) and not no_references:
+                ref_labels = rttm_to_labels(rttm_file)
+                reference = labels_to_pyannote_object(ref_labels, uniq_name=uniq_id)
+                all_reference.append([uniq_id, reference])
+            else:
+                no_references = True
+                all_reference = []
+    return all_reference, all_hypothesis
 
 def embedding_normalize(embs, use_std=False, eps=1e-10):
     """
